:recycle: MyPageProfilesection 컴포넌트화
