+++ src/pages/MyPage.tsx
@@ -1,7 +1,7 @@
 import styled, { css } from 'styled-components';
 import { GoBook } from 'react-icons/go';
-import { BsFillImageFill, BsChatLeftDots } from 'react-icons/bs';
-import { BiTrash } from 'react-icons/bi';
+import { BsChatLeftDots } from 'react-icons/bs';
+
 import { FaUserLock, FaUserAltSlash } from 'react-icons/fa';
 import { GiDiscussion } from 'react-icons/gi';
 import { IoIosArrowDown } from 'react-icons/io';
@@ -29,6 +29,7 @@ import useUpdateUsername from '../hooks/myPage/useUpdateUsername';
 import passwordValidate from '../utils/auth/passwordValidate';
 import useUpdateUserPassword from '../hooks/myPage/useUpdateUserPassword';
 import goToTop from '../utils/goToTop';
+import MyPageProfileSection from '../components/MyPage/MyPageProfileSection';
 
 function MyPage() {
   const [passwordVisible, togglePasswordVisible] = useVisibles(false);
@@ -71,10 +72,9 @@ function MyPage() {
       console.log(error);
     },
   });
-  const { mutate: updateUserInfoMutate } = useUpdateUsername({
+  const { mutate: updateUsernameMutate } = useUpdateUsername({
     onSuccess: (responceUserInfo) => {
       setUserInfo(responceUserInfo);
-      setIsUsernameChange((prev) => !prev);
     },
     onError: (error) => {
       console.log(error);
@@ -92,9 +92,6 @@ function MyPage() {
       setErrorMessage(String(error.message));
     },
   });
-
-  const [isUsernameChange, setIsUsernameChange] = useState(false);
-  const [username, setUsername] = useState(userInfo.username);
 
   function yesCallback() {
     deleteUserAccountMutate({ token });
@@ -164,19 +161,6 @@ function MyPage() {
     setModalCategory(category);
   };
 
-  const handleUsernameChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setUsername(e.target.value);
-  };
-
-  const handleUsernameChangeMode = () => {
-    setUsername(userInfo.username);
-    setIsUsernameChange((prev) => !prev);
-  };
-
-  const handleUsernameChangeButton = () => {
-    updateUserInfoMutate({ token, username: username || '' });
-  };
-
   const handleChangePasswordButton = () => {
     if (password.length > 0 && checkPassword.length > 0) {
       const { isVailed, error } = passwordValidate(password, checkPassword);
@@ -197,58 +181,11 @@ function MyPage() {
 
   return (
     <MainContainer>
-      <ProfileContianer>
-        <ImageSection>
-          <img src={userInfo.avatarUrl || undefined} alt="profile" />
-        </ImageSection>
-        <InfoSection>
-          <InfoTop>
-            {isUsernameChange ? (
-              <>
-                <UsernameInput
-                  type="text"
-                  value={username || ''}
-                  onChange={handleUsernameChange}
-                  name="changeUsername"
-                />
-                <EditButton type="button" onClick={handleUsernameChangeButton}>
-                  확인
-                </EditButton>
-                <EditButton type="button" onClick={handleUsernameChangeMode}>
-                  취소
-                </EditButton>
-              </>
-            ) : (
-              <>
-                <Username>{userInfo.username}</Username>
-                <EditButton type="button" onClick={handleUsernameChangeMode}>
-                  수정
-                </EditButton>
-              </>
-            )}
-          </InfoTop>
-          <InfoBottom>
-            <Button
-              type="button"
-              buttonType="button"
-              buttonColor="mint"
-              buttonSize="m"
-            >
-              <BsFillImageFill size={17} />
-              이미지 업로드
-            </Button>
-            <Button
-              type="button"
-              buttonType="button"
-              buttonColor="pink"
-              buttonSize="m"
-            >
-              <BiTrash size={22} />
-              이미지 제거
-            </Button>
-          </InfoBottom>
-        </InfoSection>
-      </ProfileContianer>
+      <MyPageProfileSection
+        token={token}
+        userInfo={userInfo}
+        updateUsernameMutate={updateUsernameMutate}
+      />
       <IndexContainer>
         <IndexBar>
           <IndexBarTitle>
@@ -439,60 +376,10 @@ const DeleteAccountIcon = styled(FaUserAltSlash)`
   color: var(--black);
 `;
 
-const ProfileContianer = styled.section`
-  display: flex;
-  margin-top: 64px;
-  margin-bottom: 64px;
-`;
-
-const ImageSection = styled.section`
-  border-right: 1px solid var(--color-inputbox-line);
-
-  img {
-    width: 128px;
-    border-radius: 50%;
-    margin-right: 27px;
-  }
-`;
-
-const InfoSection = styled.section`
-  margin-left: 27px;
-`;
-
-const InfoTop = styled.div`
-  display: flex;
-  align-items: flex-end;
-  margin-top: 6px;
-  margin-bottom: 28px;
-`;
-
-const usernameCss = css`
-  font-size: var(--font-size-xxl);
-  font-weight: 600;
-  margin-bottom: 4px;
-`;
-
-const Username = styled.span`
-  ${usernameCss}
-`;
-
-const UsernameInput = styled.input`
-  ${usernameCss}
-  border: 1px solid var(--color-borderbox-line);
-  &:focus {
-    outline: auto;
-  }
-`;
-
 const EditButton = styled.button`
   font-size: var(--font-size-m);
   margin-left: 19px;
   padding: 0;
-`;
-
-const InfoBottom = styled.div`
-  display: flex;
-  gap: 32px;
 `;
 
 const IndexContainer = styled.section`
